Refactor UI modules into dedicated package
