+++ src/vasoanalyzer/ui/main_window.py
@@ -39,7 +39,6 @@ from PyQt5.QtWidgets import (
     QStackedWidget,
     QUndoStack,
     QUndoView,
-    QUndoCommand,
 )
 
 from PyQt5.QtGui import QPixmap, QImage, QIcon
@@ -56,30 +55,14 @@ from vasoanalyzer.theme_manager import (
     apply_light_theme,
     css_rgba_to_mpl,
 )
+from vasoanalyzer.ui.dialogs.axis_settings_dialog import AxisSettingsDialog
+from vasoanalyzer.ui.dialogs.plot_style_dialog import PlotStyleDialog
+from vasoanalyzer.ui.dialogs.subplot_layout_dialog import SubplotLayoutDialog
+from vasoanalyzer.ui.commands import ReplaceEventCommand
 
 log = logging.getLogger(__name__)
 
-
-PREVIOUS_PLOT_PATH = os.path.join(
-    os.path.expanduser("~"), ".vasoanalyzer_last_plot.pickle"
-)
-DEFAULT_STYLE = dict(
-    axis_font_size=16,
-    axis_font_family="Arial",
-    axis_bold=True,
-    axis_italic=False,
-    tick_font_size=12,
-    event_font_size=12,
-    event_font_family="Arial",
-    event_bold=False,
-    event_italic=False,
-    pin_font_size=10,
-    pin_font_family="Arial",
-    pin_bold=False,
-    pin_italic=False,
-    pin_size=6,
-    line_width=2,
-)
+from .constants import PREVIOUS_PLOT_PATH, DEFAULT_STYLE
 
 
 # [B] ========================= MAIN CLASS DEFINITION ================================
@@ -2561,531 +2544,3 @@ class VasoAnalyzerApp(QMainWindow):
 
 
 
-# [L] ========================= AxisSettingsDialog =========================
-from PyQt5.QtWidgets import (
-    QDialog,
-    QVBoxLayout,
-    QHBoxLayout,
-    QFormLayout,
-    QPushButton,
-    QLineEdit,
-    QCheckBox,
-    QLabel,
-)
-
-
-class AxisSettingsDialog(QDialog):
-    def __init__(self, parent, ax, canvas):
-        super().__init__(parent)
-        self.ax = ax
-        self.canvas = canvas
-        self.setWindowTitle("Axis Settings")
-
-        layout = QVBoxLayout(self)
-        form = QFormLayout()
-
-        self.xmin = QLineEdit(str(ax.get_xlim()[0]))
-        self.xmax = QLineEdit(str(ax.get_xlim()[1]))
-        self.ymin = QLineEdit(str(ax.get_ylim()[0]))
-        self.ymax = QLineEdit(str(ax.get_ylim()[1]))
-        self.xlabel = QLineEdit(ax.get_xlabel())
-        self.ylabel = QLineEdit(ax.get_ylabel())
-        self.grid_chk = QCheckBox("Show Grid")
-        self.grid_chk.setChecked(any(line.get_visible() for line in ax.get_xgridlines()))
-
-        form.addRow("X Min:", self.xmin)
-        form.addRow("X Max:", self.xmax)
-        form.addRow("Y Min:", self.ymin)
-        form.addRow("Y Max:", self.ymax)
-        form.addRow("X Title:", self.xlabel)
-        form.addRow("Y Title:", self.ylabel)
-        form.addRow(self.grid_chk)
-        layout.addLayout(form)
-
-        btn_row = QHBoxLayout()
-        btn_row.addStretch()
-        apply_btn = QPushButton("Apply")
-        cancel_btn = QPushButton("Cancel")
-        ok_btn = QPushButton("OK")
-        btn_row.addWidget(apply_btn)
-        btn_row.addWidget(cancel_btn)
-        btn_row.addWidget(ok_btn)
-        layout.addLayout(btn_row)
-
-        apply_btn.clicked.connect(self.apply)
-        cancel_btn.clicked.connect(self.reject)
-        ok_btn.clicked.connect(self.accept)
-
-    def apply(self):
-        try:
-            self.ax.set_xlim(float(self.xmin.text()), float(self.xmax.text()))
-        except ValueError:
-            pass
-        try:
-            self.ax.set_ylim(float(self.ymin.text()), float(self.ymax.text()))
-        except ValueError:
-            pass
-        self.ax.set_xlabel(self.xlabel.text())
-        self.ax.set_ylabel(self.ylabel.text())
-        self.ax.grid(self.grid_chk.isChecked())
-        self.canvas.draw_idle()
-
-    def accept(self):
-        self.apply()
-        super().accept()
-
-
-# [L] ========================= PlotStyleDialog =========================
-from PyQt5.QtWidgets import (
-    QDialog,
-    QTabWidget,
-    QWidget,
-    QVBoxLayout,
-    QHBoxLayout,
-    QGridLayout,
-    QFormLayout,
-    QPushButton,
-    QLabel,
-    QComboBox,
-    QSpinBox,
-    QDoubleSpinBox,
-    QCheckBox,
-)
-
-
-class PlotStyleDialog(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Plot Style Editor")
-        self.setMinimumWidth(400)
-        self.apply_callback = None
-
-        # Consistent styling across tabs and buttons
-        self.setStyleSheet(
-            """
-            QDialog {
-                font-family: Arial;
-                font-size: 12px;
-            }
-            QTabWidget::pane {
-                margin: 6px;
-            }
-            QFormLayout {
-                margin-top: 4px;
-                margin-bottom: 4px;
-            }
-            QPushButton {
-                min-width: 70px;
-                padding: 4px 8px;
-            }
-            """
-        )
-
-        # Tab widget
-        self.tabs = QTabWidget()
-        main_layout = QVBoxLayout(self)
-        main_layout.setContentsMargins(10, 10, 10, 10)
-        main_layout.setSpacing(8)
-        main_layout.addWidget(self.tabs)
-
-        # Bottom row: Apply All / Cancel / OK
-        btn_row = QHBoxLayout()
-        btn_row.addStretch()
-        self.apply_all_btn = QPushButton("Apply")
-        self.cancel_btn = QPushButton("Cancel")
-        self.ok_btn = QPushButton("OK")
-        btn_row.addWidget(self.apply_all_btn)
-        btn_row.addWidget(self.cancel_btn)
-        btn_row.addWidget(self.ok_btn)
-        main_layout.addLayout(btn_row)
-
-        # Connect them
-        self.apply_all_btn.clicked.connect(self.handle_apply_all)
-        self.cancel_btn.clicked.connect(self.reject)
-        self.ok_btn.clicked.connect(self.accept)
-
-        # Create each tab
-        self._init_axis_tab()
-        self._init_tick_tab()
-        self._init_event_tab()
-        self._init_pin_tab()
-        self._init_line_tab()
-
-    def handle_apply_all(self):
-        """Apply *all* settings at once."""
-        style = self.get_style()
-        if callable(self.apply_callback):
-            self.apply_callback(style)
-        else:
-            self.parent().apply_plot_style(style)
-
-    def handle_apply_tab(self, section):
-        """Apply only one section (axis/tick/event/pin/line)."""
-        style = self.get_style()
-        if callable(self.apply_callback):
-            self.apply_callback(style)
-        else:
-            parent = self.parent()
-            if section == "axis":
-                parent.ax.xaxis.label.set_fontsize(style["axis_font_size"])
-                parent.ax.xaxis.label.set_fontname(style["axis_font_family"])
-                parent.ax.xaxis.label.set_fontweight(
-                    "bold" if style["axis_bold"] else "normal"
-                )
-                parent.ax.xaxis.label.set_fontstyle(
-                    "italic" if style["axis_italic"] else "normal"
-                )
-                parent.ax.yaxis.label.set_fontsize(style["axis_font_size"])
-                parent.ax.yaxis.label.set_fontname(style["axis_font_family"])
-                parent.ax.yaxis.label.set_fontweight(
-                    "bold" if style["axis_bold"] else "normal"
-                )
-                parent.ax.yaxis.label.set_fontstyle(
-                    "italic" if style["axis_italic"] else "normal"
-                )
-
-            elif section == "tick":
-                parent.ax.tick_params(axis="x", labelsize=style["tick_font_size"])
-                parent.ax.tick_params(axis="y", labelsize=style["tick_font_size"])
-
-            elif section == "event":
-                for txt, _ in parent.event_text_objects:
-                    txt.set_fontsize(style["event_font_size"])
-                    txt.set_fontname(style["event_font_family"])
-                    txt.set_fontweight("bold" if style["event_bold"] else "normal")
-                    txt.set_fontstyle("italic" if style["event_italic"] else "normal")
-
-            elif section == "pin":
-                for marker, label in parent.pinned_points:
-                    marker.set_markersize(style["pin_size"])
-                    label.set_fontsize(style["pin_font_size"])
-                    label.set_fontname(style["pin_font_family"])
-                    label.set_fontweight("bold" if style["pin_bold"] else "normal")
-                    label.set_fontstyle("italic" if style["pin_italic"] else "normal")
-
-            elif section == "line":
-                if parent.ax.lines:
-                    parent.ax.lines[0].set_linewidth(style["line_width"])
-
-            parent.canvas.draw_idle()
-
-    def _make_section_widgets(self, section):
-        """Helper: create section's Apply/Default row."""
-        h = QHBoxLayout()
-        h.setSpacing(6)
-        h.addStretch()
-        apply_btn = QPushButton("Apply")
-        default_btn = QPushButton("Default")
-        apply_btn.clicked.connect(lambda _, sec=section: self.handle_apply_tab(sec))
-        default_btn.clicked.connect(lambda _, sec=section: self.reset_defaults(sec))
-        h.addWidget(apply_btn)
-        h.addWidget(default_btn)
-        return h
-
-    def _init_axis_tab(self):
-        tab = QWidget()
-        tab.setObjectName("axis_tab")
-        layout = QVBoxLayout(tab)
-        layout.setSpacing(6)
-        desc = QLabel("Adjust axis title fonts")
-        desc.setWordWrap(True)
-        layout.addWidget(desc)
-        form = QFormLayout()
-        self.axis_font_size = QSpinBox()
-        self.axis_font_size.setRange(6, 32)
-        self.axis_font_size.setValue(14)
-        self.axis_font_family = QComboBox()
-        self.axis_font_family.addItems(
-            ["Arial", "Helvetica", "Times New Roman", "Courier", "Verdana"]
-        )
-        self.axis_bold = QCheckBox("Bold")
-        self.axis_italic = QCheckBox("Italic")
-        form.addRow("Font Size:", self.axis_font_size)
-        form.addRow("Font Family:", self.axis_font_family)
-        form.addRow("", self.axis_bold)
-        form.addRow("", self.axis_italic)
-        layout.addLayout(form)
-        layout.addLayout(self._make_section_widgets("axis"))
-        self.tabs.addTab(tab, "Axis Titles")
-
-    def _init_tick_tab(self):
-        tab = QWidget()
-        tab.setObjectName("tick_tab")
-        layout = QVBoxLayout(tab)
-        layout.setSpacing(6)
-        desc = QLabel("Set tick label font size")
-        desc.setWordWrap(True)
-        layout.addWidget(desc)
-        form = QFormLayout()
-        self.tick_font_size = QSpinBox()
-        self.tick_font_size.setRange(6, 32)
-        self.tick_font_size.setValue(12)
-        form.addRow("Tick Font Size:", self.tick_font_size)
-        layout.addLayout(form)
-        layout.addLayout(self._make_section_widgets("tick"))
-        self.tabs.addTab(tab, "Tick Labels")
-
-    def _init_event_tab(self):
-        tab = QWidget()
-        tab.setObjectName("event_tab")
-        layout = QVBoxLayout(tab)
-        layout.setSpacing(6)
-        desc = QLabel("Customize event annotation fonts")
-        desc.setWordWrap(True)
-        layout.addWidget(desc)
-        form = QFormLayout()
-        self.event_font_size = QSpinBox()
-        self.event_font_size.setRange(6, 32)
-        self.event_font_size.setValue(10)
-        self.event_font_family = QComboBox()
-        self.event_font_family.addItems(
-            ["Arial", "Helvetica", "Times New Roman", "Courier", "Verdana"]
-        )
-        self.event_bold = QCheckBox("Bold")
-        self.event_italic = QCheckBox("Italic")
-        form.addRow("Font Size:", self.event_font_size)
-        form.addRow("Font Family:", self.event_font_family)
-        form.addRow("", self.event_bold)
-        form.addRow("", self.event_italic)
-        layout.addLayout(form)
-        layout.addLayout(self._make_section_widgets("event"))
-        self.tabs.addTab(tab, "Event Labels")
-
-    def _init_pin_tab(self):
-        tab = QWidget()
-        tab.setObjectName("pin_tab")
-        layout = QVBoxLayout(tab)
-        layout.setSpacing(6)
-        desc = QLabel("Pinned point labels and marker size")
-        desc.setWordWrap(True)
-        layout.addWidget(desc)
-        form = QFormLayout()
-        self.pin_font_size = QSpinBox()
-        self.pin_font_size.setRange(6, 32)
-        self.pin_font_size.setValue(10)
-        self.pin_font_family = QComboBox()
-        self.pin_font_family.addItems(
-            ["Arial", "Helvetica", "Times New Roman", "Courier", "Verdana"]
-        )
-        self.pin_bold = QCheckBox("Bold")
-        self.pin_italic = QCheckBox("Italic")
-        self.pin_size = QSpinBox()
-        self.pin_size.setRange(2, 20)
-        self.pin_size.setValue(6)
-        form.addRow("Font Size:", self.pin_font_size)
-        form.addRow("Font Family:", self.pin_font_family)
-        form.addRow("", self.pin_bold)
-        form.addRow("", self.pin_italic)
-        form.addRow("Marker Size:", self.pin_size)
-        layout.addLayout(form)
-        layout.addLayout(self._make_section_widgets("pin"))
-        self.tabs.addTab(tab, "Pinned Labels")
-
-    def _init_line_tab(self):
-        tab = QWidget()
-        tab.setObjectName("line_tab")
-        layout = QVBoxLayout(tab)
-        layout.setSpacing(6)
-        desc = QLabel("Trace line thickness")
-        desc.setWordWrap(True)
-        layout.addWidget(desc)
-        form = QFormLayout()
-        self.line_width = QSpinBox()
-        self.line_width.setRange(1, 10)
-        self.line_width.setValue(2)
-        form.addRow("Line Width:", self.line_width)
-        layout.addLayout(form)
-        layout.addLayout(self._make_section_widgets("line"))
-        self.tabs.addTab(tab, "Trace Style")
-
-    def reset_defaults(self, section):
-        defaults = {
-            "axis": {
-                "axis_font_size": DEFAULT_STYLE["axis_font_size"],
-                "axis_font_family": DEFAULT_STYLE["axis_font_family"],
-                "axis_bold": DEFAULT_STYLE["axis_bold"],
-                "axis_italic": DEFAULT_STYLE["axis_italic"],
-            },
-            "tick": {"tick_font_size": DEFAULT_STYLE["tick_font_size"]},
-            "event": {
-                "event_font_size": DEFAULT_STYLE["event_font_size"],
-                "event_font_family": DEFAULT_STYLE["event_font_family"],
-                "event_bold": DEFAULT_STYLE["event_bold"],
-                "event_italic": DEFAULT_STYLE["event_italic"],
-            },
-            "pin": {
-                "pin_font_size": DEFAULT_STYLE["pin_font_size"],
-                "pin_font_family": DEFAULT_STYLE["pin_font_family"],
-                "pin_bold": DEFAULT_STYLE["pin_bold"],
-                "pin_italic": DEFAULT_STYLE["pin_italic"],
-                "pin_size": DEFAULT_STYLE["pin_size"],
-            },
-            "line": {"line_width": DEFAULT_STYLE["line_width"]},
-        }
-        for attr, val in defaults[section].items():
-            widget = getattr(self, attr)
-            if isinstance(widget, QSpinBox):
-                widget.setValue(val)
-            elif isinstance(widget, QComboBox):
-                widget.setCurrentText(val)
-            elif isinstance(widget, QCheckBox):
-                widget.setChecked(val)
-
-    def get_style(self):
-        return {
-            "axis_font_size": self.axis_font_size.value(),
-            "axis_font_family": self.axis_font_family.currentText(),
-            "axis_bold": self.axis_bold.isChecked(),
-            "axis_italic": self.axis_italic.isChecked(),
-            "tick_font_size": self.tick_font_size.value(),
-            "event_font_size": self.event_font_size.value(),
-            "event_font_family": self.event_font_family.currentText(),
-            "event_bold": self.event_bold.isChecked(),
-            "event_italic": self.event_italic.isChecked(),
-            "pin_font_size": self.pin_font_size.value(),
-            "pin_font_family": self.pin_font_family.currentText(),
-            "pin_bold": self.pin_bold.isChecked(),
-            "pin_italic": self.pin_italic.isChecked(),
-            "pin_size": self.pin_size.value(),
-            "line_width": self.line_width.value(),
-        }
-
-    def set_style(self, style):
-        """Populate widgets based on a style dictionary."""
-        for key, val in style.items():
-            if not hasattr(self, key):
-                continue
-            widget = getattr(self, key)
-            if isinstance(widget, QSpinBox):
-                widget.setValue(int(val))
-            elif isinstance(widget, QComboBox):
-                idx = widget.findText(str(val))
-                if idx >= 0:
-                    widget.setCurrentIndex(idx)
-                else:
-                    widget.setCurrentText(str(val))
-            elif isinstance(widget, QCheckBox):
-                widget.setChecked(bool(val))
-
-
-class SubplotLayoutDialog(QDialog):
-    def __init__(self, parent=None, fig=None):
-        super().__init__(parent)
-        self.setWindowTitle("Subplot Layout")
-        self.fig = fig
-
-        self.params = self._get_initial_params()
-
-        layout = QVBoxLayout(self)
-        form = QGridLayout()
-        self.controls = {}
-        names = ["left", "right", "top", "bottom", "wspace", "hspace"]
-        for row, name in enumerate(names):
-            label = QLabel(name.capitalize())
-            slider = QSlider(Qt.Horizontal)
-            slider.setRange(0, 100)
-            spin = QDoubleSpinBox()
-            spin.setRange(0.0, 1.0)
-            spin.setSingleStep(0.01)
-            spin.setDecimals(2)
-            slider.setValue(int(self.params[name] * 100))
-            spin.setValue(self.params[name])
-            slider.valueChanged.connect(lambda v, s=spin: s.setValue(v / 100))
-            spin.valueChanged.connect(lambda v, s=slider: s.setValue(int(v * 100)))
-            slider.valueChanged.connect(self.update_preview)
-            spin.valueChanged.connect(self.update_preview)
-            form.addWidget(label, row, 0)
-            form.addWidget(slider, row, 1)
-            form.addWidget(spin, row, 2)
-            self.controls[name] = spin
-        layout.addLayout(form)
-
-        self.preview_fig = Figure(figsize=(2, 2), facecolor=CURRENT_THEME['window_bg'])
-        self.preview_canvas = FigureCanvas(self.preview_fig)
-        self.preview_axes = self.preview_fig.subplots(2, 2)
-        for ax in self.preview_axes.flat:
-            ax.plot([0, 1], [0, 1])
-            ax.set_xticks([])
-            ax.set_yticks([])
-        layout.addWidget(self.preview_canvas)
-
-        btn_row = QHBoxLayout()
-        btn_row.addStretch()
-        ok_btn = QPushButton("OK")
-        cancel_btn = QPushButton("Cancel")
-        btn_row.addWidget(ok_btn)
-        btn_row.addWidget(cancel_btn)
-        layout.addLayout(btn_row)
-        ok_btn.clicked.connect(self.accept)
-        cancel_btn.clicked.connect(self.reject)
-
-        self.update_preview()
-
-    def _get_initial_params(self):
-        if self.fig is not None:
-            sp = self.fig.subplotpars
-            return {
-                "left": sp.left,
-                "right": sp.right,
-                "top": sp.top,
-                "bottom": sp.bottom,
-                "wspace": sp.wspace,
-                "hspace": sp.hspace,
-            }
-        return {
-            "left": rcParams.get("figure.subplot.left", 0.125),
-            "right": rcParams.get("figure.subplot.right", 0.9),
-            "top": rcParams.get("figure.subplot.top", 0.88),
-            "bottom": rcParams.get("figure.subplot.bottom", 0.11),
-            "wspace": rcParams.get("figure.subplot.wspace", 0.2),
-            "hspace": rcParams.get("figure.subplot.hspace", 0.2),
-        }
-
-    def get_values(self):
-        return {name: ctrl.value() for name, ctrl in self.controls.items()}
-
-    def update_preview(self, *_):
-        params = self.get_values()
-        # Clamp parameters to [0, 1]
-        epsilon = 1e-3
-        for key in params:
-            params[key] = max(0.0, min(1.0, params[key]))
-
-        # Ensure left < right and bottom < top
-        if params["right"] <= params["left"]:
-            if params["left"] >= 1.0 - epsilon:
-                params["left"] = 1.0 - epsilon
-                params["right"] = 1.0
-            else:
-                params["right"] = min(1.0, params["left"] + epsilon)
-
-        if params["top"] <= params["bottom"]:
-            if params["bottom"] >= 1.0 - epsilon:
-                params["bottom"] = 1.0 - epsilon
-                params["top"] = 1.0
-            else:
-                params["top"] = min(1.0, params["bottom"] + epsilon)
-
-        self.preview_fig.subplots_adjust(**params)
-        self.preview_canvas.draw_idle()
-
-
-class ReplaceEventCommand(QUndoCommand):
-    def __init__(self, app, index, old_val, new_val):
-        super().__init__(f"Replace Event #{index}")
-        self.app = app
-        self.i = index
-        self.old = old_val
-        self.new = new_val
-
-    def redo(self):
-        lbl, t, _, frame = self.app.event_table_data[self.i]
-        self.app.event_table_data[self.i] = (lbl, t, self.new, frame)
-        self.app.populate_table()
-        self.app.auto_export_table()
-
-    def undo(self):
-        lbl, t, _, frame = self.app.event_table_data[self.i]
-        self.app.event_table_data[self.i] = (lbl, t, self.old, frame)
-        self.app.populate_table()
-        self.app.auto_export_table()
